refact ppu
